Simplify mappers in TMS
Move the lookup tables for TempSensor and FanController into their
classes. No need to have them so accessible.

Replaces LinearMap with direct calculations. It was too much abstraction
to use them.
